Refactor stream::client to client::conn
