+++ src/client/conn/transport/stream.rs
@@ -4,7 +4,7 @@ use tokio::io::AsyncWrite;
 
 use tower::Service;
 
-use crate::client::stream::Stream;
+use crate::client::conn::Stream;
 use crate::info::BraidAddr;
 use crate::info::HasConnectionInfo;
 
@@ -65,8 +65,8 @@ mod fut {
 
     use pin_project::pin_project;
 
+    use crate::client::conn::Stream;
     use crate::client::conn::{Transport, TransportStream};
-    use crate::client::stream::Stream;
     use crate::info::{BraidAddr, HasConnectionInfo};
 
     /// Future returned by `IntoStream` transports.
